+++ src/main/java/cart/config/WebConfiguration.java
@@ -1,6 +1,5 @@
 package cart.config;
 
-import cart.util.MemberArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
